requirements.txt
elasticsearch>=8.0.0,<9.0.0
pyyaml>=6.0
flask>=2.3.0
flask-cors>=4.0.0
python-dateutil>=2.8.2
ipaddress>=1.0.23
geoip2>=4.7.0
maxminddb>=2.2.0
pytz>=2023.3

